src/requirements.txt
boto3==1.10.38
celery==4.3.0
Django==2.2.8
#3-compatible
django-cors-headers==3.2.0
django-environ==0.4.5
django-extensions==2.2.5
django-storages==1.8
djangorestframework==3.11.0
drf-yasg==1.17.0
gunicorn==20.0.4
gevent==1.4.0
ipython==7.10.1
psycopg2-binary==2.8.4
sentry-sdk==0.13.5
SQLAlchemy==1.3.11
setuptools==42.0.2
django-redis==4.12.1
redis==3.5.3